+++ docs/guide/intro/history.md
@@ -1,4 +1,19 @@
+---
+outline: 2
+---
 # 更新历史
 
+`Autostore`共包括以下包：
+- `autostore`: 核心库
+- `@autostorejs/react`: 用于`React`的状态管理库
+- `@autostorejs/devtools`: 调试工具库
 
+
+
+## @autostorejs/react
+
+<!--@include: ../../../packages/react/CHANGELOG.md{3,}-->
+
+## autostore
  
+<!--@include: ../../../packages/core/CHANGELOG.md{3,}-->